// src/api/generated/report-apis/models/stockAnalysis.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Define Edge Agents API
 * LangGraph-based multi-agent system for financial market analysis
 * OpenAPI spec version: 1.0.0
 */
import type { StockAnalysisReportData } from "./stockAnalysisReportData";

/**
 * Stock analysis.
 */
export interface StockAnalysis {
  /** The id of the stock analysis. */
  id: string;
  /** The ticker of the stock. */
  ticker: string;
  /** The name of the company. */
  company_name: string;
  data: StockAnalysisReportData;
  /** The date of the analysis. */
  date: string;
}